changed multplier data type to float and fixed gold/silver digit typos

+++ component-calculator-backend/db/migrate/20210130204251_create_colors.rb
@@ -3,7 +3,7 @@ class CreateColors < ActiveRecord::Migration[6.0]
     create_table :colors do |t|
       t.string :name
       t.integer :digit
-      t.integer :multiplier
+      t.float :multiplier
       t.float :tolerance
       t.string :tolerance_letter
       t.integer :temp_coefficient
